requirements.txt
functions-framework==3.*
pillow
flask
unittest2 